[수정사항] - 코칭 마크 기능 오류 버그 수정 CoroutineScope 를 이용 해야 함

+++ app/src/main/java/com/littlefox/app/foxschool/common/CommonUtils.kt
@@ -110,7 +110,7 @@ class CommonUtils
      * @param timeMs 밀리세컨드
      * @return 시간 String Ex) HH:MM:TT
      */
-    fun getMillisecondTime(timeMs : Int) : String
+    fun getMillisecondTime(timeMs : Long) : String
     {
         val mFormatBuilder = StringBuilder()
         val mFormatter = Formatter(mFormatBuilder, Locale.getDefault())
